Implement a metric exporter
When the metric exporter is deployed together with the main test container,
it will export the test data as Prometheus metrics for collection.

This can be useful when users are running a long kbench test and want to
observer the performance of the volume over time.

For example, users can deploy multiple kbench pods and enable metric
exporter. Then user can observe the performance of the volume of each
kbench pod overtime. User can also see how performance is impacted as
the number of kbench pods is increased in the cluster

Signed-off-by: Phan Le <phan.le@suse.com>